Optimised the generation of .lts files a little bit.
The .lts files were internally transformed to a probabilistic .lts file
and subsequently saved with explicit target distributions. The
transformation step is now omitted, and transitions that have only one
target state (with probability 1) are saved without transforming the
target state to an explicit distribution.

The resulting .lts file is a little smaller and the required memory
to generate the file is slightly reduced.

Old .lts files will be read properly. New .lts files cannot be read
anymore by old tools.

+++ libraries/lts/source/liblts_lts.cpp
@@ -16,7 +16,7 @@
 #include "mcrl2/data/detail/io.h"
 #include "mcrl2/lps/multi_action.h"
 #include "mcrl2/lts/lts_lts.h"
-#include "mcrl2/lts/detail/liblts_swap_to_from_probabilistic_lts.h"
+// #include "mcrl2/lts/detail/liblts_swap_to_from_probabilistic_lts.h"
 
 namespace mcrl2
 {
@@ -32,9 +32,16 @@ static atermpp::function_symbol transition_empty_header()
   static atermpp::function_symbol tr("transition_empty_list",0);
   return tr;
 }
-static atermpp::function_symbol transition_list_header()
+
+static atermpp::function_symbol probabilistic_transition_list_header()  // This contains a transition with a probabilistic target.
 {
   static atermpp::function_symbol tr("transition_list",4);
+  return tr;
+}
+
+static atermpp::function_symbol plain_transition_list_header()  // This contains a transition with a single state as target. 
+{
+  static atermpp::function_symbol tr("plain_transition_list",4);
   return tr;
 }
 
@@ -96,16 +103,37 @@ class aterm_probabilistic_transition_list: public aterm_appl
       : aterm_appl(transition_empty_header())
     {}
 
+    // Construct a probabilistic transition list element. If the target probabilistic state only contains
+    // a single state, store it as a plain_transition_element. 
     aterm_probabilistic_transition_list(const std::size_t source, 
                                         const std::size_t label, 
                                         const probabilistic_lts_lts_t::probabilistic_state_t& target,
                                         const aterm_probabilistic_transition_list& next_transition)
-      : aterm_appl(transition_list_header(), 
+      : aterm_appl((target.size()==1)?plain_transition_list_header():probabilistic_transition_list_header(), 
                    aterm_int(source),
                    aterm_int(label), 
-                   state_probability_list(target),
+                   ((target.size()==1)?
+                           static_cast<aterm>(aterm_int(target.begin()->state())):
+                           static_cast<aterm>(state_probability_list(target))),
                    next_transition)
     {} 
+
+    aterm_probabilistic_transition_list(const std::size_t source, 
+                                        const std::size_t label, 
+                                        const std::size_t target,
+                                        const aterm_probabilistic_transition_list& next_transition)
+      : aterm_appl(plain_transition_list_header(), 
+                   aterm_int(source),
+                   aterm_int(label), 
+                   aterm_int(target),
+                   next_transition)
+    {} 
+
+    bool is_probabilistic_transition()
+    {
+      assert(function()==probabilistic_transition_list_header() || function()==plain_transition_list_header());
+      return function()==probabilistic_transition_list_header();
+    }
 
     std::size_t source() const
     {
@@ -117,9 +145,19 @@ class aterm_probabilistic_transition_list: public aterm_appl
       return (atermpp::down_cast<aterm_int>((*this)[1]).value());
     }
 
-    probabilistic_lts_lts_t::probabilistic_state_t target() const
+    probabilistic_lts_lts_t::probabilistic_state_t probabilistic_target() const
     {
+      if (function()==plain_transition_list_header())
+      {
+        return probabilistic_lts_lts_t::probabilistic_state_t(atermpp::down_cast<aterm_int>((*this)[2]).value());
+      } 
+      assert(function()==probabilistic_transition_list_header());
       return aterm_list_to_probabilistic_state(atermpp::down_cast<aterm_list>((*this)[2]));
+    }
+
+    std::size_t plain_target() const
+    {
+      return atermpp::down_cast<aterm_int>((*this)[2]).value();
     }
 
     const aterm_probabilistic_transition_list& next() const 
@@ -162,6 +200,26 @@ class aterm_labelled_transition_system: public atermpp::aterm_appl
                   )
     {}
 
+    aterm_labelled_transition_system(
+               const lts_lts_t& ts,
+               const aterm_probabilistic_transition_list& transitions,
+               const state_labels_t& state_label_list,
+               const action_labels_t& action_label_list)
+      : aterm_appl(lts_header(),
+                   aterm_appl(meta_data_header(),
+                              data::detail::data_specification_to_aterm(ts.data()),
+                              ts.process_parameters(),
+                              ts.action_label_declarations(),
+                              aterm_appl(num_of_states_labels_and_initial_state(),
+                                         aterm_int(ts.num_states()),
+                                         aterm_int(ts.num_action_labels()),
+                                         state_probability_list(probabilistic_lts_lts_t::probabilistic_state_t(ts.initial_state())))),
+                   transitions,
+                   state_label_list,
+                   action_label_list
+                  )
+    {}
+
     // \brief add_index() adds a unique index to some term types, such as variables, to access data about them 
     //        quickly. When loading a term, these indices must first be added before a term can be used in the toolset.
     void add_indices()
@@ -173,22 +231,34 @@ class aterm_labelled_transition_system: public atermpp::aterm_appl
       
       (*this)=aterm(); // Clear this aterm. It is often huge, and in this way the list of transitions can be garbage collected,
                        // while being transformed. 
-      //
+      
       // Add indices to the transitions. This list is often too long to fit on the stack.
       // Therefore, it is done using this ad_hoc routine. Note that the list is reverted. 
+
       std::unordered_map<atermpp::aterm_appl, atermpp::aterm> cache;
       aterm_appl new_trans=aterm_appl(transition_empty_header());
       while (trans.function()!= transition_empty_header())
       {
         assert(trans.size()>3);
-        new_trans=aterm_appl(transition_list_header(),
+        if (trans.function()==probabilistic_transition_list_header())
+        {
+          new_trans=aterm_appl(probabilistic_transition_list_header(),
                              trans[0],
                              trans[1],
                              data::detail::add_index(trans[2],cache),
                              new_trans);
+        }
+        else
+        {
+          new_trans=aterm_appl(plain_transition_list_header(),
+                             trans[0],
+                             trans[1],
+                             trans[2],
+                             new_trans);
+        }
         trans=trans.next();
       }
-
+      
       *this = aterm_appl(lts_header(),
                              data::detail::add_index(md,cache),
                              new_trans,
@@ -197,6 +267,8 @@ class aterm_labelled_transition_system: public atermpp::aterm_appl
     }
 
     /// \brief Remove indices from dedicated terms such as variables and process names. 
+    //  Note that this code can be optimised, as in non probabilistic transition systems
+    //  the code for add indices can be skipped on transition lists. 
     void remove_indices()
     {
       aterm_appl md=meta_data();
@@ -207,19 +279,31 @@ class aterm_labelled_transition_system: public atermpp::aterm_appl
       (*this)=aterm(); // Clear this aterm. It is often huge, and in this way the list of transitions can be garbage collected,
                        // while being transformed. 
       
+      std::unordered_map<atermpp::aterm_appl, atermpp::aterm> cache;
       // Remove indices from the transitions. This list is often too long to fit on the stack.
       // Therefore, it is done using this ad_hoc routine. Note that the list is reverted. 
-      std::unordered_map<atermpp::aterm_appl, atermpp::aterm> cache;
+      
       aterm_appl new_trans=aterm_appl(transition_empty_header());
       while (trans.function()!= transition_empty_header())
       {
-        new_trans=aterm_appl(transition_list_header(),
+        if (trans.function()==probabilistic_transition_list_header())
+        {
+          new_trans=aterm_appl(probabilistic_transition_list_header(),
                                  trans[0],
                                  trans[1],
                                  data::detail::remove_index(trans[2],cache),
                                  new_trans);
+        }
+        else 
+        {
+          new_trans=aterm_appl(plain_transition_list_header(),
+                                 trans[0],
+                                 trans[1],
+                                 trans[2],
+                                 new_trans);
+        }
         trans=trans.next();
-      }
+      } 
 
       *this = aterm_appl(lts_header(),
                              data::detail::remove_index(md,cache),
@@ -289,8 +373,58 @@ class aterm_labelled_transition_system: public atermpp::aterm_appl
     }
 };
 
-static void read_from_lts(probabilistic_lts_lts_t& l, const std::string& filename)
+static void read_transitions(lts_lts_t& l, const aterm_labelled_transition_system& input_lts)
 {
+  aterm_probabilistic_transition_list input_transitions=input_lts.transitions();
+
+  while (input_transitions.function()!= transition_empty_header()) 
+  {
+    assert(input_transitions.function()==probabilistic_transition_list_header() ||
+           input_transitions.function()==plain_transition_list_header()); // In this last case, there is one probabilistic target state. 
+    if (input_transitions.is_probabilistic_transition())
+    {
+      throw mcrl2::runtime_error("Trying to read a non probabilistic LTS that appears to contain a probabilistic transition.");
+    }
+    l.add_transition(transition(input_transitions.source(), input_transitions.label(), input_transitions.plain_target()));
+    input_transitions=input_transitions.next();
+  }
+}
+
+static void read_transitions(probabilistic_lts_lts_t& l, const aterm_labelled_transition_system& input_lts)
+{
+  aterm_probabilistic_transition_list input_transitions=input_lts.transitions();
+
+  while (input_transitions.function()!= transition_empty_header()) 
+  {
+    assert(input_transitions.function()==probabilistic_transition_list_header() ||
+           input_transitions.function()==plain_transition_list_header()); // In this last case, there is one probabilistic target state. 
+    const std::size_t prob_state_index=l.add_probabilistic_state(input_transitions.probabilistic_target());
+    l.add_transition(transition(input_transitions.source(), input_transitions.label(), prob_state_index));
+    input_transitions=input_transitions.next();
+  }
+}
+
+static void set_initial_state(lts_lts_t& l, const aterm_labelled_transition_system& input_lts)
+{
+  const probabilistic_lts_lts_t::probabilistic_state_t& initial_state=input_lts.initial_probabilistic_state();
+  if (initial_state.size()>1)
+  {
+    throw mcrl2::runtime_error("The initial state of the non probabilistic input lts is probabilistic.");
+  }
+  l.set_initial_state(input_lts.initial_probabilistic_state().begin()->state());
+}
+
+static void set_initial_state(probabilistic_lts_lts_t& l, const aterm_labelled_transition_system& input_lts)
+{
+  l.set_initial_probabilistic_state(input_lts.initial_probabilistic_state());
+}
+
+template <class LTS_TRANSITION_SYSTEM>     
+static void read_from_lts(LTS_TRANSITION_SYSTEM& l, const std::string& filename)
+{
+  static_assert(std::is_same<LTS_TRANSITION_SYSTEM,probabilistic_lts_lts_t>::value || 
+                std::is_same<LTS_TRANSITION_SYSTEM,lts_lts_t>::value,
+                "Function read_from_lts can only be applied to a (probabilistic) lts. ");
   aterm input;
   if (filename=="")
   {
@@ -347,8 +481,7 @@ static void read_from_lts(probabilistic_lts_lts_t& l, const std::string& filenam
     throw runtime_error("The input file " + filename + " is not in proper .lts format. There is a problem with the datatypes, process parameters and action declarations.");
   }
   
-  aterm_labelled_transition_system input_lts(input);
-  input=aterm(); // The input is a large term. We do not need it anymore. 
+  aterm_labelled_transition_system input_lts(input); // We are copying an aterm. This is cheap. 
   input_lts.add_indices();  // Add indices to certain term types, such as variables, and process/pbes names. 
   
   l.set_data(input_lts.data());
@@ -357,15 +490,9 @@ static void read_from_lts(probabilistic_lts_lts_t& l, const std::string& filenam
 
   l.set_action_label_declarations(input_lts.action_label_declarations());
   
-  aterm_probabilistic_transition_list input_transitions=input_lts.transitions();
-  while (input_transitions.function()!= transition_empty_header()) 
-  {
-    assert(input_transitions.function()==transition_list_header());
-    const std::size_t prob_state_index=l.add_probabilistic_state(input_transitions.target());
-    l.add_transition(transition(input_transitions.source(), input_transitions.label(), prob_state_index));
-    input_transitions=input_transitions.next();
-  }
-  
+  // reading transition differs for an lts_lts and a probabilistic_lts.
+  read_transitions(l,input_lts);
+
   if (input_lts.get_state_labels().size()==0)
   {
     l.set_num_states(input_lts.num_states());
@@ -396,24 +523,66 @@ static void read_from_lts(probabilistic_lts_lts_t& l, const std::string& filenam
       }
     }
   }
-  l.set_initial_probabilistic_state(input_lts.initial_probabilistic_state());
+  set_initial_state(l, input_lts);
 }
 
-static void write_to_lts(const probabilistic_lts_lts_t& l, const std::string& filename)
+aterm_probabilistic_transition_list make_transition_list(const lts_lts_t& l)
 {
-  aterm_probabilistic_transition_list transitions;
+  aterm_probabilistic_transition_list resulting_transitions;
   
   for(std::vector<transition>::const_reverse_iterator i=l.get_transitions().rbegin();
                 i!=l.get_transitions().rend(); ++i)
   {
     // The transitions are stored in reverse order.
-    transitions=aterm_probabilistic_transition_list(
+    resulting_transitions=aterm_probabilistic_transition_list(
                                 i->from(), 
                                 l.apply_hidden_label_map(i->label()), 
-                                l.probabilistic_state(i->to()),
-                                transitions);
+                                i->to(),
+                                resulting_transitions);
   }
+  return resulting_transitions;
+}
 
+aterm_probabilistic_transition_list make_transition_list(const probabilistic_lts_lts_t& l)
+{
+  aterm_probabilistic_transition_list resulting_transitions;
+  
+  for(std::vector<transition>::const_reverse_iterator i=l.get_transitions().rbegin();
+                i!=l.get_transitions().rend(); ++i)
+  {
+    const probabilistic_lts_lts_t::probabilistic_state_t& probabilistic_state = l.probabilistic_state(i->to());
+    // The transitions are stored in reverse order.
+    if (probabilistic_state.size()==1)
+    {
+      // The target is a single state with probability 1. 
+      resulting_transitions=aterm_probabilistic_transition_list(
+                                i->from(), 
+                                l.apply_hidden_label_map(i->label()), 
+                                probabilistic_state.begin()->state(),
+                                resulting_transitions);
+    }
+    else
+    {
+      // The target is a probabilistic state with at least two plain states. 
+      resulting_transitions=aterm_probabilistic_transition_list(
+                                i->from(), 
+                                l.apply_hidden_label_map(i->label()), 
+                                probabilistic_state,
+                                resulting_transitions);
+    }
+  }
+  return resulting_transitions;
+}
+
+template <class LTS_TRANSITION_SYSTEM>     
+static void write_to_lts(const LTS_TRANSITION_SYSTEM& l, const std::string& filename)
+{
+  static_assert(std::is_same<LTS_TRANSITION_SYSTEM,probabilistic_lts_lts_t>::value || 
+                std::is_same<LTS_TRANSITION_SYSTEM,lts_lts_t>::value,
+                "Function read_from_lts can only be applied to a (probabilistic) lts. ");
+  
+  // This is different for an probabilistic lts and a non probabilistic lts. 
+  const aterm_probabilistic_transition_list transitions=make_transition_list(l);  
   state_labels_t state_label_list;
   if (l.has_state_info())
   { for(std::size_t i=l.num_state_labels(); i>0;)
@@ -471,36 +640,26 @@ static void write_to_lts(const probabilistic_lts_lts_t& l, const std::string& fi
 
 void probabilistic_lts_lts_t::save(const std::string& filename) const
 {
-  mCRL2log(log::verbose) << "Starting to save file " << filename << "\n";
+  mCRL2log(log::verbose) << "Starting to save a probabilistic lts to the file " << filename << ".\n";
+  detail::write_to_lts(*this,filename);
+}
+
+void lts_lts_t::save(std::string const& filename) const
+{
+  mCRL2log(log::verbose) << "Starting to save an lts to the file " << filename << ".\n";
   detail::write_to_lts(*this,filename);
 }
 
 void probabilistic_lts_lts_t::load(const std::string& filename)
 {
-  mCRL2log(log::verbose) << "Starting to load file " << filename << "\n";
+  mCRL2log(log::verbose) << "Starting to load a probabilistic lts from the file " << filename << ".\n";
   detail::read_from_lts(*this,filename);
 }
 
 void lts_lts_t::load(const std::string& filename)
 {
-  probabilistic_lts_lts_t l;
-  l.load(filename);
-  detail::swap_to_non_probabilistic_lts
-             <state_label_lts,
-              action_label_lts,
-              probabilistic_state<std::size_t, lps::probabilistic_data_expression>,
-              detail::lts_lts_base>(l,*this);
-}
-
-void lts_lts_t::save(std::string const& filename) const
-{
-  probabilistic_lts_lts_t l;
-  detail::translate_to_probabilistic_lts
-            <state_label_lts,
-             action_label_lts,
-             probabilistic_state<std::size_t, lps::probabilistic_data_expression>,
-             detail::lts_lts_base >(*this,l);
-  l.save(filename);
+  mCRL2log(log::verbose) << "Starting to load an lts from the file " << filename << ".\n";
+  detail::read_from_lts(*this,filename);
 }
